onchain uri set, uri populated to propery

+++ src/Players.sol
@@ -6,35 +6,22 @@ contract PlayerS {
     mapping(bytes => bool) public usernameExists;
     mapping(address => bool) public alreadyRegistered;
 
-    mapping(uint => Property) public properties;
+    mapping(uint256 => Property) public properties;
 
     struct Property {
         string name;
-        uint id;
-        uint price;
+        uint256 id;
+        uint256 price;
         address owner;
-        uint noOfTimesSold;
+        uint256 noOfTimesSold;
         bool isMortgaged;
     }
 
-    event PropertySold(uint propertyId, address owner, uint price);
-    event RentPaid(
-        address tenant,
-        address landlord,
-        uint rentPrice,
-        string property
-    );
-    event PropertyMortgaged(
-        uint propertyId,
-        uint mortgageAmount,
-        address owner
-    );
+    event PropertySold(uint256 propertyId, address owner, uint256 price);
+    event RentPaid(address tenant, address landlord, uint256 rentPrice, string property);
+    event PropertyMortgaged(uint256 propertyId, uint256 mortgageAmount, address owner);
 
-    event PropertyListedForSale(
-        uint propertyId,
-        uint propertyPrice,
-        address owner
-    );
+    event PropertyListedForSale(uint256 propertyId, uint256 propertyPrice, address owner);
 
     constructor() {}
 
@@ -48,10 +35,7 @@ contract PlayerS {
      *     @notice this function reads from the state to check if the converted lowercase username already exist to avoid duplicacy.
      *     @notice this function emits an event when a player is registered.
      */
-    function registerPlayer(
-        address playerAddress,
-        string memory username
-    ) external {
+    function registerPlayer(address playerAddress, string memory username) external {
         require(!alreadyRegistered[playerAddress], "player already registered");
 
         bytes memory _usernameBytes = convertToLowerCase(username);
@@ -93,20 +77,16 @@ contract PlayerS {
      *     @dev player should only be able to buy a property if they are not bankrupt.
      *     @dev player should only be able to buy a property if it should owned by the bank
      */
-
     function buyProperty(uint256 propertyId) external payable {
         Property storage property = properties[propertyId];
 
-        require(
-            msg.value == property.price,
-            "Insufficient Ether to buy property"
-        );
+        require(msg.value == property.price, "Insufficient Ether to buy property");
         require(property.owner != msg.sender, "You already own the property");
 
         // If previously sold, transfer funds to the current owner
         if (property.noOfTimesSold > 0) {
             // require(property.owner != address(0), "Invalid current owner");
-            (bool success, ) = property.owner.call{value: property.price}("");
+            (bool success,) = property.owner.call{value: property.price}("");
             require(success, "Transfer failed");
         }
 
@@ -128,14 +108,8 @@ contract PlayerS {
     function sellProperty(uint256 propertyId) external {
         Property storage property = properties[propertyId];
 
-        require(
-            property.owner == msg.sender,
-            "You are not the owner of this property"
-        );
-        require(
-            !property.isMortgaged,
-            "Property is mortgaged and cannot be sold"
-        );
+        require(property.owner == msg.sender, "You are not the owner of this property");
+        require(!property.isMortgaged, "Property is mortgaged and cannot be sold");
 
         emit PropertyListedForSale(propertyId, property.price, msg.sender);
     }
@@ -152,14 +126,9 @@ contract PlayerS {
     function rentProperty(uint256 propertyId) external {
         Property storage property = properties[propertyId];
         require(property.owner != address(0), "Invalid current owner");
-        (bool success, ) = property.owner.call{value: property.price}("");
+        (bool success,) = property.owner.call{value: property.price}("");
         require(success, "Transfer failed");
-        emit RentPaid(
-            msg.sender,
-            property.owner,
-            property.price,
-            property.name
-        );
+        emit RentPaid(msg.sender, property.owner, property.price, property.name);
     }
 
     /**
@@ -178,14 +147,11 @@ contract PlayerS {
     function mortgageProperty(uint256 propertyId) external {
         Property storage property = properties[propertyId];
 
-        require(
-            property.owner == msg.sender,
-            "You are not the owner of this property"
-        );
+        require(property.owner == msg.sender, "You are not the owner of this property");
         require(!property.isMortgaged, "Property is already mortgaged");
 
         property.isMortgaged = true;
-        uint mortgageAmount = property.price / 2;
+        uint256 mortgageAmount = property.price / 2;
         // Transfer funds to the contract
         payable(address(this)).transfer(mortgageAmount);
 
@@ -196,10 +162,7 @@ contract PlayerS {
     function releaseMortgage(uint256 propertyId) external payable {
         Property storage property = properties[propertyId];
 
-        require(
-            property.owner == msg.sender,
-            "You are not the owner of this property"
-        );
+        require(property.owner == msg.sender, "You are not the owner of this property");
         require(property.isMortgaged, "Property is not mortgaged");
         require(msg.value > 0, "Payment must be greater than zero");
 
@@ -212,22 +175,13 @@ contract PlayerS {
 
     function upgradeProperty(uint256 propertyId) external {}
 
-    function convertToLowerCase(
-        string memory username
-    ) private pure returns (bytes memory) {
+    function convertToLowerCase(string memory username) private pure returns (bytes memory) {
         bytes memory recievedUsernameBytes = bytes(username);
-        bytes memory convertedUsernameBytes = new bytes(
-            recievedUsernameBytes.length
-        );
+        bytes memory convertedUsernameBytes = new bytes(recievedUsernameBytes.length);
 
         for (uint256 i = 0; i < recievedUsernameBytes.length; i++) {
-            if (
-                (uint8(recievedUsernameBytes[i]) >= 65) &&
-                (uint8(recievedUsernameBytes[i]) <= 90)
-            ) {
-                convertedUsernameBytes[i] = bytes1(
-                    uint8(recievedUsernameBytes[i]) + 32
-                );
+            if ((uint8(recievedUsernameBytes[i]) >= 65) && (uint8(recievedUsernameBytes[i]) <= 90)) {
+                convertedUsernameBytes[i] = bytes1(uint8(recievedUsernameBytes[i]) + 32);
             } else {
                 convertedUsernameBytes[i] = recievedUsernameBytes[i];
             }
